+++ src/main/java/space/space_spring/controller/ChatRoomController.java
@@ -5,9 +5,9 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import space.space_spring.argument_resolver.jwtLogin.JwtLoginAuth;
-import space.space_spring.dto.chat.CreateChatRoomRequest;
-import space.space_spring.dto.chat.CreateChatRoomResponse;
-import space.space_spring.dto.chat.ReadChatRoomResponse;
+import space.space_spring.dto.chat.request.CreateChatRoomRequest;
+import space.space_spring.dto.chat.response.CreateChatRoomResponse;
+import space.space_spring.dto.chat.response.ReadChatRoomResponse;
 import space.space_spring.exception.ChatException;
 import space.space_spring.response.BaseResponse;
 import space.space_spring.service.ChatRoomService;
