+++ src/main/java/space/space_spring/dto/chat/request/ChatTestRequest.java
@@ -1,4 +1,4 @@
-package space.space_spring.dto.chat;
+package space.space_spring.dto.chat.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
