+++ src/main/java/space/space_spring/dto/chat/request/CreateChatRoomRequest.java
@@ -1,4 +1,4 @@
-package space.space_spring.dto.chat;
+package space.space_spring.dto.chat.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
