+++ src/main/java/space/space_spring/service/ChatRoomService.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Service;
 import space.space_spring.dao.UserDao;
 import space.space_spring.dao.chat.ChatRoomDao;
 import space.space_spring.dao.chat.UserChatRoomDao;
-import space.space_spring.dto.chat.ChatRoomResponse;
-import space.space_spring.dto.chat.CreateChatRoomRequest;
-import space.space_spring.dto.chat.CreateChatRoomResponse;
-import space.space_spring.dto.chat.ReadChatRoomResponse;
+import space.space_spring.dto.chat.response.ChatRoomResponse;
+import space.space_spring.dto.chat.request.CreateChatRoomRequest;
+import space.space_spring.dto.chat.response.CreateChatRoomResponse;
+import space.space_spring.dto.chat.response.ReadChatRoomResponse;
 import space.space_spring.entity.ChatRoom;
 import space.space_spring.entity.Space;
 import space.space_spring.entity.User;
